+++ Types/MarkX/set_Input.ps1
@@ -1,0 +1,7 @@
+param(
+[PSObject[]]$InputObject
+)
+
+$this | Add-Member NoteProperty '#Input' $InputObject -Force
+
+$this.Sync()